scripts/requirements.txt
redis==4.5.4
requests==2.31.0
PyYAML==6.0.1
GitPython==3.1.31
docker==6.1.2
kubernetes==26.1.0
python-terraform==0.10.1
github3.py==4.0.1
psutil==5.9.5
boto3==1.34.0
pytest==7.4.3
psycopg2-binary==2.9.9
aiohttp==3.9.1
safety==2.3.5
bandit==1.7.5
locust==2.17.0
prometheus-client==0.19.0